% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayer-matrix.R
\name{bayer}
\alias{bayer}
\title{Bayer Matrix}
\usage{
bayer(n)
}
\arguments{
\item{n}{matrix size (square matrix of side length 2^n)}
}
\value{
A matrix
}
\description{
Returns a Bayer matrix of side length 2^n. Not exported, not intended for use by the user.
}
\details{
My code related to ordered dithering is based on my own web reseach, using mainly the two articles
  \itemize{
    \item https://gist.github.com/MehdiNS/bd41bbc6db780c9409157d35d331ac80
    \item http://www.sumsar.net/blog/2019/01/image-dithering-in-r/
    }
  It might not be exactly right from a 'purist' point of view - but it seems to give the effect I am looking
  for in this application - based purely on visual aesthetics
}
