% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest-colour.R
\name{closest_colour}
\alias{closest_colour}
\title{Closest colour}
\usage{
closest_colour(r, g, b, tp_r, tp_g, tp_b, target_palette)
}
\arguments{
\item{r}{actual r (length 1)}

\item{g}{actual g (length 1)}

\item{b}{actual b (length 1)}

\item{tp_r}{target palette reds (vector)}

\item{tp_g}{target palette greens (vector)}

\item{tp_b}{target palette blues (vector)}

\item{target_palette}{target palette vector of colours (string)}
}
\description{
Not exported - not for use by the user
  For computing the nearest target palette colour to a given colour
}
