% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dither.R
\name{dither}
\alias{dither}
\title{Produce an image with ordered dithering}
\usage{
dither(
  img,
  res = 200,
  scale = NULL,
  target_palette = "c216",
  r = 1/8,
  bayer_size = 3,
  dither = "ordered",
  original = FALSE
)
}
\arguments{
\item{img}{path or URL to image}

\item{res}{horizontal resolution of output image in pixels (default = 200)
This is only used if scale = NULL (which is the default behavior). The reason for this is to try and
prevent very large images being processed by accident (large images will be slow to dither)}

\item{scale}{scaling percentage (default = NULL).
If scale is not NULL, it overrides the value of res. Full size = 100, half size = 50 etc.}

\item{target_palette}{A target palette of colours for the output image. One of
\itemize{
  \item "c216" - 216 colours spaced uniformly in RGB (default)
  \item "c64" - 64 colours spaced uniformly in RGB
  \item A vector of colours generated by the user}}

\item{r}{spread in color space for target palette (default = 1/8). This default is set to look 'OK' by eye. It is probably the factor that will
most significantly change the appearance of the dithering. Higher numbers will give 'broader' more dithering and lower numbers will give 'tighter' less dithering}

\item{bayer_size}{bayer matrix size (square matrix of side length 2^bayer_size) (default = 3)}

\item{dither}{dither type, one of
\itemize{
  \item "ordered" - Ordered dithering using a bayer matrix (default)
  \item "diffusion" - Floyd/Steinberg error diffusion dithering from \code{magick::image_map()}
  \item "none" - No dithering}}

\item{original}{return the original image (resized to \code{res})}
}
\description{
An implementation of ordered or error diffusion dithering
}
