% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayer-matrix.R
\name{rep_mat}
\alias{rep_mat}
\title{Repeat a Bayer matrix to fill rows and columns}
\usage{
rep_mat(mat, nrow_out, ncol_out)
}
\arguments{
\item{mat}{Bayer matrix}

\item{nrow_out}{Number of rows}

\item{ncol_out}{NUmber of columns}
}
\description{
Not exported. Not intended for use by the user.
}
\details{
My code related to ordered dithering is based on my own web reseach, using mainly the two articles
  \itemize{
    \item https://gist.github.com/MehdiNS/bd41bbc6db780c9409157d35d331ac80
    \item http://www.sumsar.net/blog/2019/01/image-dithering-in-r/
    }
  It might not be exactly right from a 'purist' point of view - but it seems to give the effect I am looking
  for in this application - based purely on visual aesthetics
}
