% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dither.R
\name{dither}
\alias{dither}
\title{Produce an image with ordered dithering}
\usage{
dither(
  img,
  res = 200,
  scale = NULL,
  target_palette = "extract",
  tree_depth = 0,
  n = 16,
  bayer_size = 3,
  seed = NULL,
  dither = "ordered",
  original = FALSE
)
}
\arguments{
\item{img}{path or URL to image}

\item{res}{Horixontal resolution of output image in pixels (default = 200)
This is only used if scale = NULL (which is the default behaviour). The reason for this is to try and
prevent very large images being processed by accident (large images will be slow to dither)}

\item{scale}{Scaling percentage (default = NULL).
If scale is not NULL, it overides the value of res. Full size = 100, half size = 50 etc.}

\item{target_palette}{A target palette of colours for the output image. One of
\itemize{
  \item 'extract' (default) = A colour palette of length n is extracted from the image
  \item 'greyscale' = A greyscale palette of length n is used
  \item A character vector of colours = The vector of colours is used as is
  }}

\item{tree_depth}{depth of the quantization color classification tree when \code{target_palette = "extract"}
Values of 0 or 1 allow selection of the optimal tree depth for the color reduction algorithm.
Values between 2 and 8 may be used to manually adjust the tree depth.}

\item{n}{Number of oclours to be used (default = 16) if target palette is set to 'extract' or 'greyscale'}

\item{bayer_size}{bayer matrix size (square matrix of side length 2^bayer_size)}

\item{dither}{One of
\itemize{
  \item "ordered" - Ordered dithering using a bayer matrix (default)
  \item "diffusion" - Floyd/Steinberg error diffusion dithering
  \item "none" - No dithering
  }}
}
\description{
An implementation of ordered image dithering using the bayer matrix
}
